--- src/vocana/__init__.py
@@ -1,2 +1,0 @@
-from ._core import *
-from .sdk import ObjectRefDescriptor
